feat(home): Añadir el componente Eventos y limpiar el componente Hero.
Extraer el contenido relacionado con los eventos de Hero al nuevo componente Eventos para mejorar la modularidad y la facilidad de mantenimiento. Eliminar las importaciones no utilizadas y la lógica de carrusel de Hero.
